requirements.txt
django>=4.2.6, <4.3
djangorestframework>=3.14,<3.15
flake8>=3.9.2,<3.10
mssql-django
drf-spectacular>=0.26.5, <0.27